requirements.txt
# Core
streamlit>=1.32.0
python-dotenv>=1.0.0

# LangChain
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1

# Vector Database
chromadb>=0.4.22

# ML/AI
torch>=2.2.0
transformers>=4.36.2
sentence-transformers>=2.2.2
huggingface_hub>=0.20.3

# LLM related
llama-cpp-python>=0.2.39

# Networking & Web processing
requests>=2.31.0
beautifulsoup4>=4.12.2

# Performance optimizations
pydantic>=2.5.0
ujson>=5.8.0
msgpack>=1.0.7

# Additional utilities
tqdm>=4.66.1
scikit-learn>=1.3.2
pandas>=2.1.1
numpy>=1.26.0